new blog

+++ myprojects/myproject/blog/views.py
@@ -7,7 +7,7 @@ from django.contrib.auth.decorators import login_required
 
 @login_required
 def post_list(request):
-    posts = Post.objects.filter(published_date__lte=timezone.now()).order_by('published_date')
+    posts = Post.objects.filter(created_date__lte=timezone.now()).order_by('created_date')
     return render(request, 'blog/post_list.html', {'posts': posts})
 
 @login_required
@@ -22,7 +22,7 @@ def post_new(request):
         if form.is_valid():
             post = form.save(commit=False)
             post.author = request.user
-            post.published_date = timezone.now()
+            post.created_date = timezone.now()
             post.save()
             return redirect('post_detail', pk=post.pk)
     else:
@@ -37,17 +37,20 @@ def post_edit(request, pk):
         if form.is_valid():
             post = form.save(commit=False)
             post.author = request.user
-            post.published_date = timezone.now()
+    
             post.save()
             return redirect('post_detail', pk=post.pk)
     else:
         form = PostForm(instance=post)
     return render(request, 'blog/post_edit.html', {'form': form})
 
-def post_remove(request, pk):
-	post = get_object_or_404(Post, pk=pk)
-	post.delete()
-	return redirect('post_list')
+@login_required
+def post_remove(request, pk):  
+    post = get_object_or_404(Post, pk=pk)
+    #item = Post.objects.get(pk=post.id)
+    if request.user == post.author:
+        post.delete()
+    return redirect('post_list')        
 
 def add_comment_to_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
@@ -56,8 +59,10 @@ def add_comment_to_post(request, pk):
         if form.is_valid():
             comment = form.save(commit=False)
             comment.post = post
+            comment.author = request.user
             comment.save()
             return redirect('post_detail', pk=post.pk)
     else:
         form = CommentForm()
+
     return render(request, 'blog/add_comment_to_post.html', {'form': form})
